pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "nuclei_segmentation"
version = "0.1.0"
description = "Deep Learning methods for detecting cells in histopathological images"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    { name = "Krspin Musiol", email = "kryspin.musiol@gmail.com" },
]
dependencies = [
    "torch",
    "torchvision",
    "numpy",
    "pandas",
    "scipy",
    "scikit-image",
    "scikit-learn",
    "matplotlib",
    "seaborn",
    "wandb",
    "opencv-python",
    "albumentations",
    "dvc",
    "pytorch-lightning",
    "hydra-core",
    "segmentation-models-pytorch",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov",
    "black",
    "isort",
    "flake8",
    "pre-commit",
]

[tool.pytest.ini_options]
addopts = "-ra -q"
testpaths = [
    "tests",
]
python_files = ["test_*.py"]

[tool.black]
line-length = 100
target-version = ['py39']
include = '\.pyi?$'

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3

[tool.flake8]
max-line-length = 100
extend-ignore = "E203"